fix build On NPCBOT

+++ src/server/game/AI/NpcBots/bot_ai.h
@@ -21,7 +21,7 @@ class TeleportFinishEvent;
 class AwaitStateRemovalEvent;
 
 enum CombatRating : uint8;
-enum EnchantmentSlot : uint16;
+enum EnchantmentSlot : uint8;
 enum GossipOptionIcon : uint8;
 enum MeleeHitOutcome : uint8;
